pyproject.toml
[project]
name = "pecai"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Stepan Savchenko", email = "stephan.savchenko@gmail.com" }
]
dependencies = [
    "langchain-core>=0.1.40",
    "requests>=2.31.0",
    "pydantic>=2.6.4",
    "langchain>=0.1.14",
    "sentence-transformers>=2.6.1",
    "neo4j>=5.19.0",
    "pymystem3>=0.2.0",
    "spacy>=3.7.4",
    "pymorphy3>=2.0.1",
    "python-telegram-bot>=21.0.1",
    "pyjwt>=2.8.0",
    "cryptography>=42.0.5",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
virtual = true
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pecai"]
